+++ llmsat/run_sim.py
@@ -1,6 +1,5 @@
 import krpc
-import os
-import subprocess
+
 from pathlib import Path
 from decouple import config
 from llmsat.components.obc import OBC
@@ -15,45 +14,24 @@ from langchain.tools import tool
 from langchain.chat_models import ChatOpenAI
 from langchain.schema import AIMessage, HumanMessage, SystemMessage
 from langchain.tools import Tool
+from langchain.agents import react
+import utils
 
 CHECKPOINT_NAME = "checkpoint"
 
 
-def is_ksp_running():
-    try:
-        # List processes and grep for KSP
-        output = subprocess.check_output("tasklist", shell=True).decode("utf-8")
-        return "KSP" in output
-    except:
-        return False
-
-
-def launch_ksp(path: Path):
-    subprocess.Popen([path], cwd=os.path.dirname(path))
-
-
-def load_checkpoint(name: str, space_center):
-    """Load the given game save state"""
-    try:
-        space_center.load(name)
-    except ValueError as e:
-        raise ValueError(
-            f"No checkpoint named '{name}.sfs' was found. Please create one"
-        )
-
-
 if __name__ == "__main__":
-    if not is_ksp_running():
+    if not utils.is_ksp_running():
         ksp_path = Path(str(config("KSP_PATH")))
         print(f"Launching KSP from '{ksp_path}'...")
-        launch_ksp(path=ksp_path)
+        utils.launch_ksp(path=ksp_path)
 
     input("Press any key once the KSP save is loaded to continue...")
 
     print("Connecting to KSP...")
     connection = krpc.connect(name="Simulator")
     print(f"Loading '{CHECKPOINT_NAME}.sfs' checkpoint...")
-    load_checkpoint(name=CHECKPOINT_NAME, space_center=connection.space_center)
+    utils.load_checkpoint(name=CHECKPOINT_NAME, space_center=connection.space_center)
 
     science = connection.space_center.science
     print(science)
@@ -63,7 +41,7 @@ if __name__ == "__main__":
 
     # initialize agent
     KEY = str(config("OPENAI", cast=str))
-    llm = OpenAI(openai_api_key=KEY)
+    llm = ChatOpenAI(openai_api_key=KEY, model="gpt-4-0613")
     print(llm)
 
     system_message = SystemMessage(
