+++ llmsat/utils.py
@@ -1,0 +1,28 @@
+"""Game-related utilities"""
+
+import os
+import subprocess
+from pathlib import Path
+
+
+def is_ksp_running():
+    try:
+        # List processes and grep for KSP
+        output = subprocess.check_output("tasklist", shell=True).decode("utf-8")
+        return "KSP" in output
+    except:
+        return False
+
+
+def launch_ksp(path: Path):
+    subprocess.Popen([path], cwd=os.path.dirname(path))
+
+
+def load_checkpoint(name: str, space_center):
+    """Load the given game save state"""
+    try:
+        space_center.load(name)
+    except ValueError as e:
+        raise ValueError(
+            f"No checkpoint named '{name}.sfs' was found. Please create one"
+        )
